changes based on PR feedback
