+++ src/content/app/tools/blast/components/blast-species-selector/BlastSpeciesSelector.scss
@@ -45,9 +45,9 @@
   margin: 0 0.6rem 0 0.6rem;
 }
 
-.speciesSelectorPlaceholder {
+.speciesSelectorContainer {
   width: 860px;
-  height: 684px;
+  height: 96%;
   background-color: $light-grey;
   border: 1px solid $grey;
   padding: 10px;
@@ -82,6 +82,8 @@
   thead th {
     border-bottom: 2px solid $blue;
     padding: 6px 10px 6px 15px;
+    font-weight: 300;
+    font-size: 11px;
   }
 
   tbody td {
@@ -103,7 +105,6 @@
 }
 
 .assemblyCol {
-  font-weight: 400;
   font-size: 11px;
 }
 
@@ -111,13 +112,7 @@
   width: 235px;
 }
 
-.selectCol {
+thead th.selectCol {
   text-align: center;
-  font-weight: 400;
-  font-size: 11px;
-}
-
-.speciesDetails {
-  font-weight: 300;
-  font-size: 11px;
+  font-weight: $bold;
 }
